backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-genai>=1.0.0
feedparser>=6.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
slack-bolt>=1.20.0
aiosmtplib>=3.0.0
apscheduler>=3.10.0
jinja2>=3.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
tzdata>=2024.1
